+++ pages/side-panel/src/side-panel.tsx
@@ -1,8 +1,8 @@
 import '@src/side-panel.css';
 import { withErrorBoundary, withSuspense } from '@extension/shared';
 import { SignedIn, SignedOut } from '@clerk/chrome-extension';
-import Chatroom from './pages/chatroom';
-import Signin from './pages/signin';
+import Chatroom from './pages/chat-room';
+import Signin from './pages/sign-in';
 import { SelectedTabsProvider } from './hooks/use-selected-tabs';
 const SidePanel = () => {
   return (
